pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "supervisor"
version = "0.1.0"
description = "AI CLI Orchestrator - Treats AI CLIs as stateless workers to prevent context dilution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Auto Coder Team" }
]
keywords = ["ai", "orchestrator", "cli", "claude", "codex", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "click>=8.1",
    "rich>=13.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
supervisor = "supervisor.cli:main"

[tool.hatch.build.targets.sdist]
include = ["supervisor"]

[tool.hatch.build.targets.wheel]
packages = ["supervisor"]

[tool.hatch.version]
path = "supervisor/__init__.py"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=supervisor --cov-report=term-missing"
